requirements.txt
streamlit
pandas
openpyxl
matplotlib
Pillow